requirements.txt
brotli==1.0.9
cryptography==3.4.6
Cython==0.29.22
Flask==1.1.2
ipaddr==2.2.0
keyring==22.3.0
mock==4.0.3
mypy_extensions==0.4.3
numpy==1.20.1
ordereddict==1.1
protobuf==3.15.3
pyOpenSSL==20.0.1
simplejson==3.17.2
wincertstore==0.2
python-dotenv